feat(ws): STOMP 확장 1차 구현 – DTO/세션관리/컨트롤러/Quartz 연동
- backend
  - WebSocketMessage<T>, AuctionStatusChangeData DTO 추가 (한글 주석)
  - WebSocketSessionManager 인메모리 구독 관리(메인/카테고리/상세) 구현
  - AuctionWebSocketController 추가
    - /app/subscribe/main|category|auction 구독 처리 및 /queue/subscription-confirmed 응답
    - /app/ping → /topic/pong 하트비트
    - NEW_AUCTION, AUCTION_STATUS_CHANGE 타겟팅 브로드캐스트 헬퍼
  - AuctionStartJob/AuctionEndJob에 상태 변경 브로드캐스트 연동
- rules
  - backEnd/shrimp-rules.md, AuctionService_FE/shrimp-rules.md 추가/갱신
  - “모든 코드 주석 한글 작성” 규칙 반영

영향
- 폴링 제거를 위한 WS 기반 푸시 준비 완료 (초기 로딩 제외)
- 이후 경매 생성 시 NEW_AUCTION 브로드캐스트 및 FE 훅 연동 예정

+++ backEnd/src/main/java/org/example/bidflow/global/ws/AuctionStatusChangeData.java
@@ -22,7 +22,7 @@ public class AuctionStatusChangeData {
 	private String newStatus; // e.g., NEW_AUCTION 대상 상태 또는 변경된 상태 필드
 	private Long categoryId;
 	private LocalDateTime eventTime;
-	private Long startPrice;
+	private Integer startPrice;
 	private LocalDateTime endTime;
 	private Map<String, Object> additionalInfo;
 }
